# docs/swagger/gym-platform.yaml
...
openapi: 3.0.3
info:
  title: Gym Management Backend API
  description: |-
    Custom OpenAPI documentation for the Gym Management microservices. The platform
    manages authentication, trainees, trainers, assignments, training sessions, and
    synchronized monthly trainer workload.
  contact:
    name: Gym Management API Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
...
servers:
- url: http://localhost:8080
  description: Gym Management microservice
- url: http://localhost:8080/workload
  description: Trainer Workload microservice
...
paths:
  "/api/v1/users/update/{userId}/profile":
    put:
      tags:
      - Gym Management User Profile Controller
      summary: Update user profile by user ID
      operationId: updateUserProfile
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserProfileResponseDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainers/update/{username}":
    put:
      tags:
      - Gym Management Trainer Controller
      summary: Update a trainer profile by username
      operationId: updateTrainerProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateTrainerProfileRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/users/create":
    post:
      tags:
      - Gym Management User Profile Controller
      summary: Create a new user profile
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProfileRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserSummaryResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainings/add":
    post:
      tags:
      - Gym Management Training Controller
      summary: Add a new training session
      operationId: addTraining
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AddTrainingRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainings":
    post:
      tags:
      - Gym Management Training Controller
      summary: Add a new training session
      operationId: addTraining_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AddTrainingRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainers/register":
    post:
      tags:
      - Gym Management Trainer Controller
      summary: Register a new trainer
      operationId: registerTrainer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TrainerRegistrationRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/trainers/{username}/deactivate":
    patch:
      tags:
      - Gym Management Trainer Controller
      summary: Deactivate a trainer profile by username
      operationId: deactivateTrainerProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainers/{username}/activate":
    patch:
      tags:
      - Gym Management Trainer Controller
      summary: Activate a trainer profile by username
      operationId: activateTrainerProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/users/{userId}/profile":
    get:
      tags:
      - Gym Management User Profile Controller
      summary: Get user profile by user ID
      operationId: getUserProfile
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/users/summary":
    get:
      tags:
      - Gym Management User Profile Controller
      summary: Get user by user ID
      operationId: getUserById
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserSummaryResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/users/profiles":
    get:
      tags:
      - Gym Management User Profile Controller
      summary: Get all user profiles
      operationId: getAllUserProfiles
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/users/all":
    get:
      tags:
      - Gym Management User Profile Controller
      summary: Get all users
      operationId: getAllUsers
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserSummaryResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainings/types":
    get:
      tags:
      - Gym Management Training Controller
      summary: Get training type information based on the provided training type name
      operationId: getTrainingType
      parameters:
      - name: trainingType
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TrainingTypeResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainings/trainer/{username}":
    get:
      tags:
      - Gym Management Training Controller
      summary: Get training sessions for a trainer based on various filters
      operationId: getTrainerTrainings
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      - name: periodFrom
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: periodTo
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: traineeName
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TrainingResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainings/trainee/{username}":
    get:
      tags:
      - Gym Management Training Controller
      summary: Get training sessions for a trainee based on various filters
      operationId: getTraineeTrainings
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      - name: periodFrom
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: periodTo
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: trainerName
        in: query
        required: false
        schema:
          type: string
      - name: trainingType
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TrainingResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainers/training-types":
    get:
      tags:
      - Gym Management Trainer Controller
      summary: Get all available training types
      operationId: getTrainingTypes
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TrainingTypeResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainers/profile/{username}":
    get:
      tags:
      - Gym Management Trainer Controller
      summary: Get a trainer profile by username
      operationId: getTrainerProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/users/delete/{userId}":
    delete:
      tags:
      - Gym Management User Profile Controller
      summary: Delete user by user ID
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
      security:
      - Bearer Authentication: []
  "/api/v1/trainings/{trainingId}":
    delete:
      tags:
      - Gym Management Training Controller
      summary: Cancel a planned training session
      operationId: deleteTraining
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
...
components:
  schemas:
    UserProfileResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
        specialization:
          type: string
        profileStatus:
          type: string
          enum:
          - Active
          - Inactive
          - Suspended
          - Pending
          - Deleted
        trainers:
          type: array
          items:
            "$ref": "#/components/schemas/UserSummaryResponseDTO"
        trainees:
          type: array
          items:
            "$ref": "#/components/schemas/UserSummaryResponseDTO"
    UserSummaryResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: string
        active:
          type: boolean
        profileStatus:
          type: string
          enum:
          - Active
          - Inactive
          - Suspended
          - Pending
          - Deleted
        role:
          type: string
          enum:
          - Admin
          - Trainee
          - Trainer
    UpdateTrainerProfileRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        isActive:
          type: boolean
      required:
      - firstName
      - isActive
      - lastName
...
    MessageResponseDTO:
      type: object
      properties:
        message:
          type: string
    ProfileRequestDTO:
      type: object
      properties:
        role:
          type: string
          enum:
          - Admin
          - Trainee
          - Trainer
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        specialization:
          type: string
          enum:
          - Fitness
          - Yoga
          - Zumba
          - Stretching
          - Resistance
          - Set Later
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
          minLength: 1
      required:
      - address
      - dateOfBirth
      - firstName
      - lastName
      - role
    AddTrainingRequestDTO:
      type: object
      properties:
        traineeUsername:
          type: string
          minLength: 1
        trainerUsername:
          type: string
          minLength: 1
        trainingName:
          type: string
          minLength: 1
        trainingDate:
          type: string
          format: date
        trainingType:
          type: string
          minLength: 1
        trainingDuration:
          type: integer
          format: int32
          minimum: 1
      required:
      - traineeUsername
      - trainerUsername
      - trainingDate
      - trainingDuration
      - trainingName
      - trainingType
    TrainerRegistrationRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        specialization:
          type: string
          minLength: 1
      required:
      - firstName
      - lastName
      - specialization
    AuthResponseDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        message:
          type: string
        token:
          type: string
        tokenType:
          type: string
        role:
          type: string
...
    TrainingTypeResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trainingType:
          type: string
    TrainingResponseDTO:
      type: object
      properties:
        trainingId:
          type: string
          format: uuid
        trainingName:
          type: string
        trainingDate:
          type: string
          format: date
        trainingType:
          type: string
        trainingDuration:
          type: integer
          format: int32
        trainerName:
          type: string
        traineeName:
          type: string
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Service JWT issued with the shared deployment secret
      scheme: bearer
      bearerFormat: JWT